refactor(viewer): extract viewer_awards.jsx from viewer.jsx
Pure file split, no behavior change. Step 6 of mp-pxxc.
