+++ web-mobile/index.html
@@ -91,6 +91,10 @@
          extracted from viewer.jsx (mp-pxxc step 5).
          Must load after viewer_match.js and before viewer.js. -->
     <script type="module" src="/dist/viewer_standings.js?v=5"></script>
+    <!-- Awards components (deriveAwards, AwardsView, FightingSpiritSection, AllWinnersView, …),
+         extracted from viewer.jsx (mp-pxxc step 6).
+         Must load after viewer_standings.js and before viewer.js. -->
+    <script type="module" src="/dist/viewer_awards.js?v=5"></script>
     <script type="module" src="/dist/viewer.js?v=5"></script>
     <script type="module" src="/dist/admin_helpers.js?v=5"></script>
     <script type="module" src="/dist/admin_shell.js?v=5"></script>
